requirements.txt
streamlit==1.25.0
pandas==2.1.4
python-dotenv==1.0.0
openai==0.28.0